Added table of Lesson times on contact page with loop to insert the days
Fixed some validation errors on the contact page. There are still some errors from the "for" values you have on there, check that out. Don't want to mess with it because I am not sure exactly what you were doing with them. 

Added a new table at the bottom to display the time we are open for lessons. Used a while loop function to insert the days to the text that I added which was the time. It also uses an array so I guess that gets two for one. 

Next I am going to polish up my function the main page some then I'll creat another loop somewhere else.