Moved all JavaScript to js page
I have moved all the javascript to the js page to have all the javascript in a centralized area. It all works as it did before, however it is generating the error "Uncaught TypeError: Cannot read property 'addEventListener' of null". Still working on a fix to this error before I merge this project branch with the master.