Couldn't merge stacy's branch so just copied from it to master.
The branch was too different to automatically merge the two, but I just copied the relevent code from it to the master. Also made some visual and JavaScript additions. Form is good now. Also checked all the code for JavaScript and/or valdiation errors, it is all good. As far as I can tell right now with this we are done. Feel free to make more changes however, and/or test the code further.